ref(meme): create folders and split files
