+++ client/caption/services/save-caption.js
@@ -1,9 +1,13 @@
-import { apiService as api } from "../api/api.service.js";
+import { apiService as api } from "../../api/api.service.js";
 
-export const saveCaption = (tText,bText) => {
+const saveCaption = (tText,bText) => {
     api.save(`create-caption`,{topText: tText, bottomText: bText}
     ).then(()=> {
         console.log('caption saved'+ tText + bText)
     }).catch(() => console.error('caption not saved, please try again'))
 
 }
+
+
+export default saveCaption;
+
