+++ client/index.html
@@ -5,24 +5,22 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Meme Generator App using MENJ</title>
-    <script type="module" src="./index.js"></script>
 
     <!-- Bootstrap CSS -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
 
     <!-- CSS Stylesheet -->
     <link rel="stylesheet" href="./shared/global.css">
-    <link rel="stylesheet" href="./meme/meme.css">
+    <link rel="stylesheet" href="./meme/components/meme.css">
 
-    <script src="./index.js" type="module"></script>
+    <script src="./meme/index.js" type="module"></script>
 
     <!-- Google Fonts Link -->
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Chau+Philomene+One&family=Roboto&display=swap" rel="stylesheet">
 
-    <title>meme-generator</title>
+    <title>Meme Generator</title>
 </head>
 
 <body>
