--- client/index.js
@@ -1,4 +1,0 @@
-import makeAMeme from '../client/meme/create-meme.js';
-
-
-makeAMeme();
