+++ client/meme/index.js
@@ -1,0 +1,4 @@
+import makeAMeme from '../meme/services/create-meme.js';
+
+
+makeAMeme();
